+++ src/store/json.rs
@@ -117,7 +117,9 @@ mod tests {
         project
             .add_concept("Child".into(), Some(ConceptId(1)), Some("desc".into()))
             .unwrap();
-        project.add_task("Do stuff".into(), Some("details".into()), Some(2.5), None);
+        project
+            .add_task("Do stuff".into(), Some("details".into()), Some(2.5), None)
+            .unwrap();
 
         let file = NamedTempFile::new().unwrap();
         save(file.path(), &project).unwrap();